+++ CHANGELOG.md
@@ -37,9 +37,9 @@ The types of changes are:
 * Wrap up the email connector - it sends an email with erasure instructions as part of request execution [#1246](https://github.com/ethyca/fidesops/pull/1246)
 * Mapping Vault environment variables in docker-compose.yml [#1275](https://github.com/ethyca/fidesops/pull/1275)
 * Foundations for a new "manual_webhook" connector type [#1267](https://github.com/ethyca/fidesops/pull/1267)
-* Adds Fixture to Datadog Tests [1269](https://github.com/ethyca/fidesops/pull/1269)
+* Data seeding for Datadog access tests [#1269](https://github.com/ethyca/fidesops/pull/1269)
 
-### Docs 
+### Docs
 
 * Fix analytics opt out environment variable name [#1170](https://github.com/ethyca/fidesops/pull/1170)
 * Added how to view a subject request history and reprocess a subject request [#1164](https://github.com/ethyca/fidesops/pull/1164)
@@ -53,6 +53,7 @@ The types of changes are:
 * Fix for pytest-asyncio bug [#1260](https://github.com/ethyca/fidesops/pull/1260)
 * Fix download link in privacy center [#1264](https://github.com/ethyca/fidesops/pull/1264)
 * Make admin ui work when volumes are mounted [#1266](https://github.com/ethyca/fidesops/pull/1266)
+* Fixed typo in enum value [#1280](https://github.com/ethyca/fidesops/pull/1280)
 
 ## [1.7.2](https://github.com/ethyca/fidesops/compare/1.7.1...1.7.2)